Scaling for RC input and change to work with HK 23U4
I applied a linear scaling to the input of the RC so it would work with
my RC controller.

RC input now is 1.4X - 600
This scales so the Low point is 1000 and the high point is 2000.

This code is only for my controller as I cant scale the output values on
my rc controller.